Refactor and enhance player and match API logic.
Refactored controller and service methods for better response handling, including status codes and detailed error messages. Updated repositories to support advanced filtering by player attributes and match associations. Added RSpec and Swagger dependencies for testing, alongside updated routes to streamline API endpoints.

+++ app/services/player_service.rb
@@ -1,25 +1,51 @@
 class PlayerService
-  def self.get_current_player(auth_id)
-    PlayerRepository.find_by_auth0_id(auth_id)
+  def self.get_player_by_auth0_id(auth_id)
+    player = PlayerRepository.find_by_auth0_id(auth_id)
+    if player.nil?
+      return { success: false, status: :not_found, message: "Player not found" }
+    end
+    { success: true, status: :ok, player: player }
   end
 
-  def self.search_players(search_query = nil)
-    players = PlayerRepository.all_players
-    players = players.where("name LIKE ?", "%#{search_query}%") if search_query.present?
-    players
+  def self.search_players(player_name)
+    PlayerRepository.all_players(player_name)
   end
 
   def self.create_player(params)
-    player = Player.new(params)
-    PlayerRepository.create(player)
+    player = PlayerRepository.find_by_auth0_id(params[:auth0_id])
+    if player
+      return { success: false, status: :conflict, message: "Player already exists" }
+    end
+    new_player = PlayerRepository.create(params)
+    { success: true, status: :created, player: new_player }
   end
 
-  def self.update_player(player, updated_params)
+  def self.update_player(player_id, updated_params)
+    player = PlayerRepository.find_by_id(player_id)
+    if player.nil?
+      return { success: false, status: :not_found, message: "Player not found" }
+    end
     player.assign_attributes(updated_params)
     PlayerRepository.update(player)
+    { success: true, status: :ok, player: player }
   end
 
-  def self.delete_player(player)
+  def self.update_player_me(auth0_id, updated_params)
+    player = PlayerRepository.find_by_auth0_id(auth0_id)
+    if player.nil?
+      return { success: false, status: :not_found, message: "Player not found" }
+    end
+    player.assign_attributes(updated_params)
+    PlayerRepository.update(player)
+    { success: true, status: :ok, player: player }
+  end
+
+  def self.delete_player(player_id)
+    player = PlayerRepository.find_by_id(player_id)
+    if player.nil?
+      return { success: false, status: :not_found, message: "Player not found" }
+    end
     PlayerRepository.destroy(player)
+    { success: true, status: :ok }
   end
 end
